Extracted complicated code into individual components for the flights table.
